backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic[email]
passlib[bcrypt]
pillow
numpy
bcrypt